lib: more deserializer work
Slowly making the deserializer match up with aoe4world's API. It's slow going sadly
because the API is very poorly documented and you need to look in several places in the
codebase to get a complete picture of what's what.

TODO: split this up into modules because it's getting unwieldy.